# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Taskaya API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks/{id}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: Get all chat messages for a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a chat message on a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /notifications/preferences:
    get:
      operationId: getNotificationPreferences
      tags: [notifications]
      summary: Get notification preferences for the current user
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
    put:
      operationId: updateNotificationPreferences
      tags: [notifications]
      summary: Update notification preferences for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferences"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
...
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Generate daily report data
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportData"
  /reports/employee:
    get:
      operationId: getEmployeeReport
      tags: [reports]
      summary: Generate employee report data
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Employee report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportData"
...
components:
  schemas:
...
    NotificationPreferences:
      type: object
      required: [reminder24h, reminder1h, reminder10m, overdue]
      properties:
        reminder24h:
          type: boolean
        reminder1h:
          type: boolean
        reminder10m:
          type: boolean
        overdue:
          type: boolean
...
    User:
      type: object
      required: [id, fullName, mobile, role, isActive, mustChangePassword, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        mobile:
          type: string
        role:
          type: string
          enum: [owner, deputy, member]
        isActive:
          type: boolean
        mustChangePassword:
          type: boolean
        groupId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Task:
      type: object
      required: [id, title, description, assigneeId, creatorId, deadline, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: integer
        assignee:
          $ref: "#/components/schemas/User"
        creatorId:
          type: integer
        creator:
          $ref: "#/components/schemas/User"
        deadline:
          type: string
        status:
          type: string
          enum: [open, completed, approved, reopened]
        reassignToId:
          type: ["integer", "null"]
        reassignTo:
          $ref: "#/components/schemas/User"
        reassignStatus:
          type: ["string", "null"]
          enum: [pending, approved, rejected, "null"]
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Attachment:
      type: object
      required: [id, taskId, filename, originalName, mimeType, size, url, createdAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        url:
          type: string
        createdAt:
          type: string
    Message:
      type: object
      required: [id, taskId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        senderId:
          type: integer
        sender:
          $ref: "#/components/schemas/User"
        content:
          type: string
        attachmentUrl:
          type: ["string", "null"]
        attachmentName:
          type: ["string", "null"]
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        attachmentUrl:
          type: ["string", "null"]
        attachmentName:
          type: ["string", "null"]
    Notification:
      type: object
      required: [id, userId, type, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [task_assigned, deadline_approaching, task_completed, task_approved, task_reopened]
        message:
          type: string
        taskId:
          type: ["integer", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
...
    ReportData:
      type: object
      required: [title, period, total, completed, approved, overdue, tasks]
      properties:
        title:
          type: string
        period:
          type: string
        total:
          type: integer
        completed:
          type: integer
        approved:
          type: integer
        overdue:
          type: integer
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        employeeName:
          type: ["string", "null"]
...